split common.storage.util into smaller pieces
All code from common.storage.util is now in
- common.storage.util, or
- common.licenses.licenses, or
- common.urls

Also, common.storage.constants was moved to common.licenses.constants
