+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/licenses/constants.py
